Refactor ApiClientErrors
