+++ fpx/Cargo.toml
@@ -8,6 +8,9 @@ repository = { workspace = true }
 
 [lib]
 
+[features]
+fpx_client = ["dep:reqwest"]
+
 [dependencies]
 anyhow = { version = "1.0", default-features = false }
 async-trait = { version = "0.1", default-features = false }
@@ -30,6 +33,12 @@ opentelemetry-proto = { version = "0.27", default-features = false, features = [
     "with-schemars",
 ] }
 prost = { version = "0.13", default-features = false }
+reqwest = { version = "0.12", optional = true, default-features = false, features = [
+    "charset",
+    "http2",
+    "rustls-tls",
+    "json",
+] }
 schemars = { workspace = true, default-features = false, features = ["derive"] }
 serde = { version = "1.0", default-features = false, features = ["derive"] }
 serde_json = { version = "1.0", default-features = false }
