+++ fpx/src/api/client.rs
@@ -4,17 +4,20 @@
 //! api module. But for now this is only used within our own code, so it is
 //! fine.
 
-use super::errors::ApiClientError;
+use super::errors::CommonError;
+use super::handlers::spans::SpanGetError;
+use super::handlers::traces::TraceGetError;
+use super::models;
+use crate::otel::HeaderMapInjector;
 use anyhow::Result;
-use fpx::api::handlers::spans::SpanGetError;
-use fpx::api::handlers::traces::TraceGetError;
-use fpx::api::models;
-use fpx::otel::HeaderMapInjector;
+use bytes::Bytes;
 use http::{HeaderMap, Method, StatusCode};
 use opentelemetry::propagation::TextMapPropagator;
 use opentelemetry_sdk::propagation::TraceContextPropagator;
 use std::error::Error;
 use std::future::Future;
+use thiserror::Error;
+use tracing::error;
 use tracing::trace;
 use tracing_opentelemetry::OpenTelemetrySpanExt;
 use url::Url;
@@ -168,6 +171,54 @@ impl ApiClient {
         let path = format!("v1/traces/{}/spans/{}", trace_id.as_ref(), span_id.as_ref());
 
         self.do_req(Method::DELETE, path, no_body).await
+    }
+}
+
+#[allow(dead_code)]
+#[derive(Debug, Error)]
+pub enum ApiClientError<E> {
+    /// This can only occur when a invalid base URL was provided.
+    #[error("An invalid URL was provided: {0}")]
+    ParseError(#[from] url::ParseError),
+
+    /// An error occurred in reqwest.
+    #[error("An error occurred while making the request: {0}")]
+    ClientError(#[from] reqwest::Error),
+
+    /// An error returned from the service. These errors are specific to the
+    /// endpoint that was called.
+    #[error(transparent)]
+    ServiceError(E),
+
+    #[error(transparent)]
+    CommonError(#[from] CommonError),
+
+    /// A response was received, but we were unable to deserialize it. The
+    /// status code and the receive body are returned.
+    #[error("API returned an unknown response: Status: {0}, Body: {1:?}")]
+    InvalidResponse(StatusCode, Bytes),
+}
+
+impl<E> ApiClientError<E>
+where
+    E: serde::de::DeserializeOwned,
+{
+    /// Try to parse the result as a ServiceError or a CommonError. If both
+    /// fail, return the status_code and body.
+    pub fn from_response(status_code: StatusCode, body: Bytes) -> Self {
+        // Try to parse the result as a ServiceError.
+        if let Ok(result) = serde_json::from_slice::<E>(&body) {
+            return ApiClientError::ServiceError(result);
+        }
+
+        // Try to parse the result as CommonError.
+        if let Ok(result) = serde_json::from_slice::<CommonError>(&body) {
+            return ApiClientError::CommonError(result);
+        }
+
+        // If both failed, return the status_code and the body for the user to
+        // debug.
+        ApiClientError::InvalidResponse(status_code, body)
     }
 }
