+++ fpx/src/api/errors.rs
@@ -1,5 +1,4 @@
 use axum::response::IntoResponse;
-use bytes::Bytes;
 use fpx_macros::ApiError;
 use http::StatusCode;
 use serde::{Deserialize, Serialize};
@@ -74,53 +73,6 @@ where
 {
     fn from(value: E) -> Self {
         ApiServerError::ServiceError(value)
-    }
-}
-
-#[allow(dead_code)]
-#[derive(Debug, Error)]
-pub enum ApiClientError<E> {
-    /// This can only occur when a invalid base URL was provided.
-    #[error("An invalid URL was provided: {0}")]
-    ParseError(#[from] url::ParseError),
-
-    // /// An error occurred in reqwest.
-    // #[error("An error occurred while making the request: {0}")]
-    // ClientError(#[from] reqwest::Error),
-    /// An error returned from the service. These errors are specific to the
-    /// endpoint that was called.
-    #[error(transparent)]
-    ServiceError(E),
-
-    #[error(transparent)]
-    CommonError(#[from] CommonError),
-
-    /// A response was received, but we were unable to deserialize it. The
-    /// status code and the receive body are returned.
-    #[error("API returned an unknown response: Status: {0}, Body: {1:?}")]
-    InvalidResponse(StatusCode, Bytes),
-}
-
-impl<E> ApiClientError<E>
-where
-    E: serde::de::DeserializeOwned,
-{
-    /// Try to parse the result as a ServiceError or a CommonError. If both
-    /// fail, return the status_code and body.
-    pub fn from_response(status_code: StatusCode, body: Bytes) -> Self {
-        // Try to parse the result as a ServiceError.
-        if let Ok(result) = serde_json::from_slice::<E>(&body) {
-            return ApiClientError::ServiceError(result);
-        }
-
-        // Try to parse the result as CommonError.
-        if let Ok(result) = serde_json::from_slice::<CommonError>(&body) {
-            return ApiClientError::CommonError(result);
-        }
-
-        // If both failed, return the status_code and the body for the user to
-        // debug.
-        ApiClientError::InvalidResponse(status_code, body)
     }
 }
